Split bam track (#1797)
bam track code reorginzation
